% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.data-data.R
\docType{data}
\name{correlated data}
\alias{correlated data}
\title{Various correlation structures}
\format{A list with 4 elements containing data.frames:
\describe{
  \item{example 1}{two columns with nonlinear wave function relationship}
  \item{example 2}{two columns with simple nonlinear relationship}
  \item{example 3}{two columns with nonlinear multi-level wave function relationship}
  \item{example 4}{4 coulmuns with firs two having linear relationship}
}}
\description{
a list object with various linear and nonlinear correlation structures
}
