% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleTest.R
\name{ensembleTest}
\alias{ensembleTest}
\title{Tests the correlation across a Bootstrap ensemble to ensure
a quasi-independent sample and uncorrelated ensemble.}
\usage{
ensembleTest(
  x,
  perm = 99,
  p = 0.05,
  sampling = c("boot", "elimination"),
  type = c("mvc", "cov", "psi", "pairwise"),
  n = NULL,
  replacement = TRUE
)
}
\arguments{
\item{x}{A matrix or data.frame of the data to test}

\item{perm}{Number of Bootstraps (eg., same as random forests ntree)}

\item{p}{Accept/Reject p-value}

\item{type}{Type of matrix similarity statistic}

\item{n}{Sample proportion}

\item{replacement}{(TRUE/FALSE) Sample with replacement}
}
\value{

}
\description{
Tests the correlation across a Bootstrap ensemble to ensure
a quasi-independent sample and uncorrelated ensemble.
}
\note{
type options are:
  \itemize{ 
  \item mcv - multivariate correlation (RV-coefficient) 
  \item cov - covariance equivalence
  \item psi - Procrustes Similarity Index 
  \item pairwise - Averaged column pairwise comparison  
  }
}
\references{
Robert, P., Escoufier, Y. (1976). A Unifying Tool for Linear Multivariate 
  Statistical Methods: The RV-Coefficient. Applied Statistics 25(3):257-265.

Smilde, AK., Kiers, HA., Bijlsma, S., Rubingh, CM., van Erk, MJ (2009)
  Matrix correlations for high-dimensional data: the modified RV-coefficient. 
  Bioinformatics 25(3): 401-5.

library(raster)
r <- stack("C:/evans/India/Chennai/data/2019_OLI8.tif")
  x <- sampleRandom(r, 5000)
}
\author{
Jeffrey S. Evans   <jeffrey_evans<at>tnc.org>
}
