% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.occurrence.threshold.R
\name{plot.occurrence.threshold}
\alias{plot.occurrence.threshold}
\title{Plot occurrence thresholds}
\usage{
\method{plot}{occurrence.threshold}(x, ...)
}
\arguments{
\item{x}{A occurrence.threshold object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot function for occurrence.threshold object
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
