% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rf.cv.R
\name{plot.rf.cv}
\alias{plot.rf.cv}
\title{Plot random forests cross-validation}
\usage{
\method{plot}{rf.cv}(x, type = "cv", stat = "kappa", ...)
}
\arguments{
\item{x}{A rf.cv object}

\item{type}{Which result to evaluate c("cv","model")}

\item{stat}{Which statistic to plot: classification: "users.accuracy", "producers.accuracy", 
"kappa", "oob", regression: "rmse", "mse", "var.exp", "mae", "mbe"}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot function for rf.cv object
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
