% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rf.modelSel.R
\name{plot.rf.modelSel}
\alias{plot.rf.modelSel}
\title{Plot random forests model selection}
\usage{
\method{plot}{rf.modelSel}(x, imp = "sel", ...)
}
\arguments{
\item{x}{A rf.modelSel object}

\item{imp}{Plot selected ("sel") or all ("all") importance used in model selection}

\item{...}{Additional arguments passed to plot}
}
\description{
Dot plot function for rf.modelSel importance values
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
