% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.significance.R
\name{plot.significance}
\alias{plot.significance}
\title{Plot random forests significance}
\usage{
\method{plot}{significance}(x, ...)
}
\arguments{
\item{x}{A significance object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot function for significance object
}
\author{
Jeffrey S. Evans  <jeffrey_evans@tnc.org>
}
