% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.accuracy.R
\name{print.accuracy}
\alias{print.accuracy}
\title{Print accuracy}
\usage{
\method{print}{accuracy}(x, ...)
}
\arguments{
\item{x}{Object of class accuracy}

\item{...}{Ignored}
}
\description{
print method for class "accuracy"
}
