% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.occurrence.threshold.R
\name{print.occurrence.threshold}
\alias{print.occurrence.threshold}
\title{Print occurrence.threshold}
\usage{
\method{print}{occurrence.threshold}(x, ...)
}
\arguments{
\item{x}{Object of class occurrence.threshold}

\item{...}{Ignored}
}
\description{
Print method for occurrence.threshold objects
}
