% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rf.ensembles.R
\name{print.rf.ensembles}
\alias{print.rf.ensembles}
\title{Print for combined random forests ensembles}
\usage{
\method{print}{rf.ensembles}(x, ...)
}
\arguments{
\item{x}{Object of class rf.ensembles}

\item{...}{Ignored}
}
\description{
print method for combined random forests ensembles
}
