% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rf.modelSel.R
\name{print.rf.modelSel}
\alias{print.rf.modelSel}
\title{Print random forests model selection}
\usage{
\method{print}{rf.modelSel}(x, ...)
}
\arguments{
\item{x}{Object of class rf.modelSel}

\item{...}{Ignored}
}
\description{
Print method for rf.modelSel objects
}
