% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.significance.R
\name{print.significance}
\alias{print.significance}
\title{Print significance}
\usage{
\method{print}{significance}(x, p = 0.05, ...)
}
\arguments{
\item{x}{Object of class significance}

\item{p}{P-value to accept or reject significance}

\item{...}{Ignored}
}
\description{
print method for class "significance"
}
