% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_proximity.R
\name{ranger_proximity}
\alias{ranger_proximity}
\title{random forest ranger proximity matrix}
\usage{
ranger_proximity(y, x, sparse = FALSE)
}
\arguments{
\item{y}{A range object with keep.inbag = TRUE}

\item{x}{Independent data used in model}

\item{sparse}{(FALSE/TRUE) Output a sparse matrix}
}
\value{
An n x n matrix where position i, j gives the proportion 
        of times observation i and j are in the same terminal 
        node across all trees. If sparse = TRUE, object is a 
        sparse matrix dsCMatrix class from the Matrix package
}
\description{
Calculates OOB proximities from ranger object
}
\examples{
library(ranger)
fit <- ranger(Species ~ ., iris, keep.inbag = TRUE)
p <- ranger_proximity(fit, iris[,-5]) 

}
\author{
Jeffrey S. Evans  <jeffrey_evans<at>tnc.org
}
