% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.significance.R
\name{rf.significance}
\alias{rf.significance}
\title{Random Forest model significance test}
\usage{
rf.significance(x, nperm = 999, randomization = 1, kappa = FALSE)
}
\arguments{
\item{x}{randomForest class object}

\item{nperm}{Number of permutations}

\item{randomization}{Fraction (0.01-1) of randomization,
default is 1 (total randomization)}

\item{kappa}{(FALSE/TRUE) In classification, use kappa rather
than percent correctly classified}
}
\value{
A list class object with the following components:
For Regression problems:
\itemize{
\item RandR.square Vector of random R-square values
\item R.square The R-square of the "true" model
\item p.value p-values of randomizations of R-square
\item ks.p.value p-value(s) evaluation of Kolmogorov-Smirnov test
\item nPerm number of permutations
\item rf.type Type of Random Forests
\item rand.frac Amortization fraction
}

For Classification problems:
\itemize{
\item RandOOB Vector of random out-of-bag (OOB) values
\item RandMaxError Maximum error of randomizations
\item test.OOB Error OOB error of the "true" model
\item test.MaxError maximum class OOB error of the "true" model
\item p.value p-value based on Mcnemar's test
\item oop.p.value p-value based on permutation of OOB error
\item nPerm Number of permutations
\item rf.type Type of Random Forests
\item rand.frac Amortization fraction
}
}
\description{
Performs significance test for classification and regression
Random Forests models.
}
\details{
If the p-value is small, it suggests a near certainty
that the difference between the two populations is
significant. alternative = c("two.sided", "less", "greater")
}
\note{
Please note that previous versions of this function required xdata and "..."
arguments that are no longer necessary. The model object is now used in
obtaining the data and arguments used in the original model
}
\examples{
\dontrun{
#### Regression
library(randomForest)
library(ranger)

  set.seed(1234)	
  data(airquality)
  airquality <- na.omit(airquality)
 
 # randomForest 
 ( rf.mdl <- randomForest(x=airquality[,2:6], y=airquality[,1]) )
   ( rf.perm <- rf.significance(rf.mdl, nperm=99) )

 # ranger
 ( rf.mdl <- ranger(x=airquality[,2:6], y=airquality[,1]) )
   ( rf.perm <- rf.significance(rf.mdl, nperm=99) )

 
#### Classification
ydata = as.factor(ifelse(airquality[,1] < 40, 0, 1))
( rf.mdl <- ranger(x = airquality[,2:6], y = ydata) )
      ( rf.perm <- rf.significance(rf.mdl, nperm=99) )
   
( rf.mdl <- randomForest(x = airquality[,2:6], y = ydata) )
     ( rf.perm <- rf.significance(rf.mdl, nperm=99) )


  set.seed(1234)	
    data(iris)
      iris$Species <- as.factor(iris$Species) 

 	
 ( rf.mdl <- randomForest(x=iris[,1:4], y=iris[,"Species"]) )
   ( rf.perm <- rf.significance(rf.mdl, nperm=99) )

 ( rf.mdl <- ranger(x=iris[,1:4], y=iris[,"Species"]) )
   ( rf.perm <- rf.significance(rf.mdl, nperm=99) )

}

}
\references{
Murphy M.A., J.S. Evans, and A.S. Storfer (2010) Quantify Bufo
boreas connectivity in Yellowstone National Park with landscape
genetics. Ecology 91:252-261

Evans J.S., M.A. Murphy, Z.A. Holden, S.A. Cushman (2011). Modeling
species distribution and change using Random Forests CH.8 in
PredictiveModeling in Landscape Ecology eds Drew, CA,
Huettmann F, Wiersma Y. Springer
}
\author{
Jeffrey S. Evans \href{mailto:jeffrey_evans@tnc.org}{jeffrey_evans@tnc.org}
}
