% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.accuracy.R
\name{summary.accuracy}
\alias{summary.accuracy}
\title{Summarizing accuracy}
\usage{
\method{summary}{accuracy}(object, ...)
}
\arguments{
\item{object}{Object of class accuracy}

\item{...}{Ignored}
}
\description{
Summary method for class "accuracy".
}
