% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.occurrence.threshold.R
\name{summary.occurrence.threshold}
\alias{summary.occurrence.threshold}
\title{Summarizing occurrence.threshold}
\usage{
\method{summary}{occurrence.threshold}(object, ...)
}
\arguments{
\item{object}{Object of occurrence.threshold}

\item{...}{Ignored}
}
\description{
Summarize occurrence.threshold
}
