% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rf.cv.R
\name{summary.rf.cv}
\alias{summary.rf.cv}
\title{Summarizing cross-validation}
\usage{
\method{summary}{rf.cv}(object, ...)
}
\arguments{
\item{object}{Object of class rf.cv}

\item{...}{Ignored}
}
\description{
Summarizing of the rf.crossValidation function
}
