% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rf.ensembles.R
\name{summary.rf.ensembles}
\alias{summary.rf.ensembles}
\title{Summary for combined random forests ensembles}
\usage{
\method{summary}{rf.ensembles}(object, ...)
}
\arguments{
\item{object}{Object of class rf.ensembles}

\item{...}{Ignored}
}
\description{
summary method for combined random forests ensembles
}
