% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rf.modelSel.R
\name{summary.rf.modelSel}
\alias{summary.rf.modelSel}
\title{Summarizing random forests model selection}
\usage{
\method{summary}{rf.modelSel}(object, ...)
}
\arguments{
\item{object}{Object of class rf.modelSel}

\item{...}{Ignored}
}
\description{
Summarizing of the rf.modelSel function
}
