% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.significance.R
\name{summary.significance}
\alias{summary.significance}
\title{Summarizing significance}
\usage{
\method{summary}{significance}(object, ...)
}
\arguments{
\item{object}{Object of class significance}

\item{...}{Ignored}
}
\description{
Summarizing of a rf.significance object
}
