% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colinear.R
\name{collinear}
\alias{collinear}
\title{Collinearity test}
\usage{
collinear(x, p = 0.85)
}
\arguments{
\item{x}{A symmetric correlation matrix}

\item{p}{The correlation cutoff (default is 0.85)}
}
\value{
Messages and a vector of correlated variables
}
\description{
Test for collinearity in data
}
\note{
Evaluation of the pairwise correlated variables to remove is accomplished through 
      calculating the mean correlations of each variable and selecting the 
      variable with higher mean.
}
\examples{
dat <- data.frame(v1=seq(0.1, 5, length=100), 
                  v2=seq(0.1, 5, length=100), 
                  v3=dnorm(runif(100)), 
				     v4=dnorm(runif(100)))

# Evaluate collinearity
( cor.vars <- collinear(cor(dat), p = 0.80) )
 			       
# Remove identified variable(s)
head( dat[,-which(names(dat) \%in\% cor.vars)] )

}
\author{
Jeffrey S. Evans  <jeffrey_evans<at>tnc.org>
}
